+++ base/v0_5_exp/schema.go
@@ -118,17 +118,15 @@ type Link struct {
 }
 
 type Luks struct {
-	Clevis     Clevis       `yaml:"clevis"`
-	Device     *string      `yaml:"device"`
-	KeyFile    Resource     `yaml:"key_file"`
-	Label      *string      `yaml:"label"`
-	Name       string       `yaml:"name"`
-	Options    []LuksOption `yaml:"options"`
-	UUID       *string      `yaml:"uuid"`
-	WipeVolume *bool        `yaml:"wipe_volume"`
+	Clevis     Clevis   `yaml:"clevis"`
+	Device     *string  `yaml:"device"`
+	KeyFile    Resource `yaml:"key_file"`
+	Label      *string  `yaml:"label"`
+	Name       string   `yaml:"name"`
+	Options    []string `yaml:"options"`
+	UUID       *string  `yaml:"uuid"`
+	WipeVolume *bool    `yaml:"wipe_volume"`
 }
-
-type LuksOption string
 
 type NodeGroup struct {
 	ID   *int    `yaml:"id"`
@@ -189,14 +187,12 @@ type Proxy struct {
 }
 
 type Raid struct {
-	Devices []Device     `yaml:"devices"`
-	Level   *string      `yaml:"level"`
-	Name    string       `yaml:"name"`
-	Options []RaidOption `yaml:"options"`
-	Spares  *int         `yaml:"spares"`
+	Devices []Device `yaml:"devices"`
+	Level   *string  `yaml:"level"`
+	Name    string   `yaml:"name"`
+	Options []string `yaml:"options"`
+	Spares  *int     `yaml:"spares"`
 }
-
-type RaidOption string
 
 type Resource struct {
 	Compression  *string      `yaml:"compression"`