+++ config/fcos/v1_5_exp/schema.go
@@ -20,9 +20,8 @@ import (
 
 type Config struct {
 	base.Config `yaml:",inline"`
-	BootDevice  BootDevice  `yaml:"boot_device"`
-	Extensions  []Extension `yaml:"extensions"`
-	Grub        Grub        `yaml:"grub"`
+	BootDevice  BootDevice `yaml:"boot_device"`
+	Grub        Grub       `yaml:"grub"`
 }
 
 type BootDevice struct {
@@ -40,10 +39,6 @@ type BootDeviceLuks struct {
 
 type BootDeviceMirror struct {
 	Devices []string `yaml:"devices"`
-}
-
-type Extension struct {
-	Name string `yaml:"name"`
 }
 
 type Grub struct {