+++ docs/release-notes.md
@@ -21,6 +21,7 @@ nav_order: 9
 
 - Drop `LuksOption` and `RaidOption` types _(Go API for fcos 1.5.0-experimental,
   flatcar 1.1.0-experimental, openshift 4.13.0-experimental)_
+- Require Go 1.18+
 
 ### Docs changes