+++ docs/release-notes.md
@@ -23,6 +23,7 @@ nav_order: 9
 
 ### Misc. changes
 
+- Drop `extensions` section _(fcos 1.5.0-exp, openshift 4.13.0-exp)_
 - Drop `LuksOption` and `RaidOption` types _(Go API for fcos 1.5.0-experimental,
   flatcar 1.1.0-experimental, openshift 4.13.0-experimental)_
 - Require Go 1.18+