+++ docs/release-notes.md
@@ -16,6 +16,7 @@ nav_order: 9
   flatcar 1.1.0-exp, openshift 4.13.0-exp)_
 - Allow specifying arbitrary LUKS open options _(fcos 1.5.0-exp,
   flatcar 1.1.0-exp, openshift 4.13.0-exp)_
+- Allow specifying user password hash _(openshift 4.13.0-exp)_
 
 ### Bug fixes