+++ docs/release-notes.md
@@ -22,7 +22,7 @@ nav_order: 9
 - Allow specifying user password hash _(openshift 4.13.0+)_
 - Support offline Tang provisioning via pre-shared advertisement _(fcos 1.5.0+, openshift 4.14.0-exp)_
 - Support local file embedding for SSH keys and systemd units _(fcos 1.5.0+,
-  flatcar 1.1.0+, openshift 4.14.0-exp, r4e 1.1.0-exp)_
+  flatcar 1.1.0+, openshift 4.14.0-exp, r4e 1.1.0+)_
 
 ### Bug fixes